+++ workers/workers/scripts/purge_staged_datasets.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import workers.utils as utils
 import workers.api as api
 from workers.config import config
-from workers.dataset import get_bundle_staged_path
+from workers.dataset import get_bundle_staged_path, is_dataset_locked_for_writes
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -27,7 +27,7 @@ def main():
         'staged_path': None
     }
     for dataset in datasets[:MAX_PURGES]:
-        locked, latest_state = utils.is_dataset_locked_for_writes(dataset)
+        locked, latest_state = is_dataset_locked_for_writes(dataset)
         if locked:
             logger.warning(f"Dataset {dataset['id']} is locked for writes. Dataset will not be purged.'")
             continue
