+++ workers/workers/tasks/archive.py
@@ -11,6 +11,7 @@ import workers.config.celeryconfig as celeryconfig
 import workers.utils as utils
 import workers.workflow_utils as wf_utils
 from workers.config import config
+from workers.dataset import is_dataset_locked_for_writes
 
 app = Celery("tasks")
 app.config_from_object(celeryconfig)
@@ -79,7 +80,7 @@ def archive(celery_task: WorkflowTask, dataset: dict, delete_local_file: bool = 
 def archive_dataset(celery_task, dataset_id, **kwargs):
     dataset = api.get_dataset(dataset_id=dataset_id, bundle=True)
 
-    locked, latest_state = utils.is_dataset_locked_for_writes(dataset)
+    locked, latest_state = is_dataset_locked_for_writes(dataset)
     if locked:
         raise Exception(f"Dataset {dataset['id']} is locked for writes. Dataset's current "
                         f"state is {latest_state}.")
