+++ workers/workers/tasks/delete.py
@@ -4,6 +4,7 @@ import workers.api as api
 import workers.config.celeryconfig as celeryconfig
 import workers.sda as sda
 import workers.utils as utils
+from workers.dataset import is_dataset_locked_for_writes
 
 app = Celery("tasks")
 app.config_from_object(celeryconfig)
@@ -12,7 +13,7 @@ app.config_from_object(celeryconfig)
 def delete_dataset(celery_task, dataset_id, **kwargs):
     dataset = api.get_dataset(dataset_id=dataset_id)
 
-    locked, latest_state = utils.is_dataset_locked_for_writes(dataset)
+    locked, latest_state = is_dataset_locked_for_writes(dataset)
     if locked:
         raise Exception(f"Dataset {dataset['id']} is locked for writes. Dataset's current "
                         f"state is {latest_state}.")
