+++ workers/workers/tasks/delete_source.py
@@ -5,7 +5,8 @@ from celery import Celery
 
 import workers.api as api
 import workers.config.celeryconfig as celeryconfig
-import workers.utils as utils
+from workers.dataset import is_dataset_locked_for_writes
+
 
 app = Celery("tasks")
 app.config_from_object(celeryconfig)
@@ -14,7 +15,7 @@ app.config_from_object(celeryconfig)
 def delete_source(celery_task, dataset_id, **kwargs):
     dataset = api.get_dataset(dataset_id=dataset_id)
 
-    locked, latest_state = utils.is_dataset_locked_for_writes(dataset)
+    locked, latest_state = is_dataset_locked_for_writes(dataset)
     if locked:
         raise Exception(f"Dataset {dataset['id']} is locked for writes. Dataset's current "
                         f"state is {latest_state}.")
