+++ workers/workers/tasks/illumina_download.py
@@ -8,7 +8,7 @@ import workers.api as api
 import workers.config.celeryconfig as celeryconfig
 from workers import illumina
 from workers.config import config
-import workers.utils as utils
+from workers.dataset import is_dataset_locked_for_writes
 
 app = Celery("tasks")
 app.config_from_object(celeryconfig)
@@ -28,7 +28,7 @@ def download_recent_datasets(celery_task: WorkflowTask, download_dir: Path, n_da
 def download_illumina_dataset(celery_task, dataset_id, **kwargs):
     dataset = api.get_dataset(dataset_id=dataset_id)
 
-    locked, latest_state = utils.is_dataset_locked_for_writes(dataset)
+    locked, latest_state = is_dataset_locked_for_writes(dataset)
     if locked:
         raise Exception(f"Dataset {dataset['id']} is locked for writes. Dataset's current "
                         f"state is {latest_state}.")
