+++ workers/workers/tasks/stage.py
@@ -10,7 +10,7 @@ from sca_rhythm import WorkflowTask
 
 import workers.api as api
 import workers.utils as utils
-from workers.config import config
+from workers.dataset import is_dataset_locked_for_writes
 import workers.config.celeryconfig as celeryconfig
 import workers.workflow_utils as wf_utils
 from workers.dataset import compute_staging_path
@@ -100,7 +100,7 @@ def stage(celery_task: WorkflowTask, dataset: dict) -> (str, str):
 def stage_dataset(celery_task, dataset_id, **kwargs):
     dataset = api.get_dataset(dataset_id=dataset_id, bundle=True)
 
-    locked, latest_state = utils.is_dataset_locked_for_writes(dataset)
+    locked, latest_state = is_dataset_locked_for_writes(dataset)
     if locked:
         raise Exception(f"Dataset {dataset['id']} is locked for writes. Dataset's current "
                         f"state is {latest_state}.")
