+++ workers/workers/tasks/validate.py
@@ -9,6 +9,7 @@ import workers.api as api
 import workers.config.celeryconfig as celeryconfig
 import workers.utils as utils
 from workers import exceptions as exc
+from workers.dataset import is_dataset_locked_for_writes
 
 app = Celery("tasks")
 app.config_from_object(celeryconfig)
@@ -33,7 +34,7 @@ def check_files(celery_task: WorkflowTask, dataset_dir: Path, files_metadata: li
 def validate_dataset(celery_task, dataset_id, **kwargs):
     dataset = api.get_dataset(dataset_id=dataset_id, files=True)
 
-    locked, latest_state = utils.is_dataset_locked_for_writes(dataset)
+    locked, latest_state = is_dataset_locked_for_writes(dataset)
     if locked:
         raise Exception(f"Dataset {dataset['id']} is locked for writes. Dataset's current "
                         f"state is {latest_state}.")
